+++ src/main/java/finalmission/shopkeeper/application/ShopkeeperService.java
@@ -1,6 +1,6 @@
 package finalmission.shopkeeper.application;
 
-import finalmission.shopkeeper.application.in.dto.CreateShopkeeper;
+import finalmission.shopkeeper.application.in.dto.SignupShopkeeper;
 import finalmission.shopkeeper.application.out.ShopkeeperRepository;
 import finalmission.shopkeeper.domain.Shopkeeper;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +15,7 @@ public class ShopkeeperService {
     public final ShopkeeperRepository shopkeeperRepository;
 
     @Transactional
-    public void signUp(final CreateShopkeeper command) {
+    public void signUp(final SignupShopkeeper command) {
         shopkeeperRepository.save(
                 Shopkeeper.create(command.name())
         );
