+++ src/main/java/finalmission/shopkeeper/application/in/dto/SignupShopkeeper.java
@@ -1,4 +1,4 @@
 package finalmission.shopkeeper.application.in.dto;
 
-public record CreateShopkeeper(String name) {
+public record SignupShopkeeper(String name) {
 }
